create step1.spec.ts and create-template.spec.ts
